debug

+++ SaleManages/SaleManages/SaleManages/DAO/ObjCustomerDAO.cs
@@ -111,13 +111,18 @@ namespace SaleManages.DAO
         {
             System.Windows.Forms.Form f = System.Windows.Forms.Application.OpenForms["_frmSalesManage"];
             string tbmaso = ((_frmSalesManage)f).tbCodeKh.Text;
-            string DeleteQuery = "DELETE FROM KHACHHANG WHERE MAKH = '"+tbmaso+"'";
-            int result = DataProvider.Instance.ExecuteNonQuery(DeleteQuery);
-            if (result > 0)
+            string CheckQuery ="SELECT * FROM HOADON " +
+                "WHERE MAKH = '"+tbmaso+"' ";
+            DataTable dt = DataProvider.Instance.ExecuteQuery(CheckQuery);
+            if (dt.Rows.Count == 0)
             {
-                MessageBox.Show("Khách hàng đã bị xoá,bấm xem để xem dữ liệu mới", "Thông báo", MessageBoxButtons.OK);
+                string DeleteQuery = "DELETE FROM KHACHHANG WHERE MAKH = '" + tbmaso + "'";
+                int result = DataProvider.Instance.ExecuteNonQuery(DeleteQuery);
+                if (result > 0)
+                {
+                    MessageBox.Show("Khách hàng đã bị xoá,bấm xem để xem dữ liệu mới", "Thông báo", MessageBoxButtons.OK);
+                }
             }
-            LoadCustomerData();
         }
         public void Update()
         {
